Reducing technical debt, general improves.
Reducing technical debt. Database parameter changed to respect db architecture. Prepared audit entity to support employees leaving company.

+++ spring-boot-backend/src/main/java/com/auditorias/springboot/backend/model/Empresa.java
@@ -2,15 +2,15 @@ package com.auditorias.springboot.backend.model;
 
 public class Empresa {
 
-	private int id_company;
+	private Long id_company;
 	private String name_company;
 	private Long id_user_boss;
 
-	public int getId_company() {
+	public Long getId_company() {
 		return id_company;
 	}
 
-	public void setId_company(int id_company) {
+	public void setId_company(Long id_company) {
 		this.id_company = id_company;
 	}
